Scratch that. Reconfig is for It3. Delete LeagueStats class
